Extract CalibrationBlockRender to components/hardware-sim.
