+++ components/src/hardware-sim/Labware/index.ts
@@ -1,4 +1,5 @@
 export * from './labwareInternals/index'
+export * from './CalibrationBlockRender'
 export * from './LabwareRender'
 export * from './Labware'
